bin: fix issues with passing variables around
setting variables in one script and assuming they will
be passed on if we didn't rfork en is a bad idea.
in fact, the variables will exist with the correct
content in /env, but will be empty for rc, until
something makes it re-read /env.
forking a new rc works, subshells work sometimes.
it seemed somewhat logical to assume that namespace
extends to rc variables, and it sort of does... but
nope.  stupid to begin with?